[Bug] Incorrect output for object arrays
When serializing an `object[]` the type information is erased because
there is no guarantee that the `ElementType` of the array is
homogeneous.

There is no real "fix" for this except to assume that we see an object
array we should blindly assume that the user wants to serialize this as
jsonapi.